task 3 fixed

+++ 0x15-file_io/3-cp.c
@@ -7,7 +7,7 @@
  */
 int main(int argc, char *argv[])
 {
-	ssize_t wr, from, to, rd;
+	int wr, from, to, rd;
 	int c = 0, c1 = 0;
 	char buff[1024];
